refactor: separate useUserJoinedSpace and useSpaceAndEvent
